+++ AirRater/app/src/main/java/com/example/brandon/airrater/ExperienceItem.java
@@ -7,10 +7,10 @@ public class ExperienceItem {
 
     public String businessName;
     public int numUsers;
-    public int numStars;
+    public Double numStars;
     public int position;
 
-    public ExperienceItem(String businessName, int numUsers, int numStars, int position)
+    public ExperienceItem(String businessName, int numUsers, Double numStars, int position)
     {
         this.businessName = businessName;
         this.numUsers = numUsers;
